requirements.txt
flask
flask-sqlalchemy
flask-migrate
anthropic
python-dotenv
